Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
xvar <- c("sex", "married", "age", "havejob", "educ", "political.afl",
"rural", "region", "fin.intermdiaries", "fin.knowldge", "income")
BchMk.MLM <- MLM_Model(sample_data, c(xvar, "networth"), yvar )
#> + Fold01: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 364.317387
#> iter  20 value 350.753768
#> iter  30 value 348.444136
#> iter  40 value 348.181058
#> final  value 348.178037 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 366.434031
#> iter  20 value 353.088211
#> iter  30 value 351.658889
#> iter  40 value 351.523920
#> final  value 351.522126 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 364.319525
#> iter  20 value 350.757043
#> iter  30 value 348.447503
#> iter  40 value 348.184515
#> final  value 348.181498 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 492.002887
#> iter  20 value 378.086846
#> iter  30 value 361.436815
#> iter  40 value 348.894798
#> iter  50 value 348.482864
#> final  value 348.477989 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 504.496632
#> iter  20 value 386.479663
#> iter  30 value 364.091569
#> iter  40 value 352.995781
#> iter  50 value 352.744397
#> final  value 352.738697 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 492.016367
#> iter  20 value 378.098122
#> iter  30 value 361.427281
#> iter  40 value 348.867995
#> iter  50 value 348.487672
#> final  value 348.482721 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 495.675932
#> iter  20 value 380.896733
#> iter  30 value 354.550296
#> iter  40 value 346.585212
#> iter  50 value 346.364625
#> final  value 346.364506 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 506.464626
#> iter  20 value 380.806674
#> iter  30 value 354.117154
#> iter  40 value 350.490000
#> iter  50 value 350.423600
#> iter  50 value 350.423597
#> iter  50 value 350.423597
#> final  value 350.423597 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 495.687719
#> iter  20 value 380.905010
#> iter  30 value 354.553294
#> iter  40 value 346.589456
#> iter  50 value 346.369058
#> final  value 346.368945 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 481.097305
#> iter  20 value 366.585522
#> iter  30 value 352.339796
#> iter  40 value 348.945767
#> iter  50 value 348.844031
#> final  value 348.843999 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 491.395945
#> iter  20 value 378.396305
#> iter  30 value 360.604675
#> iter  40 value 353.998627
#> iter  50 value 353.456140
#> final  value 353.455998 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 481.108664
#> iter  20 value 366.601032
#> iter  30 value 352.362729
#> iter  40 value 348.966816
#> iter  50 value 348.865035
#> iter  60 value 348.860601
#> final  value 348.849958 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 476.063523
#> iter  20 value 371.366262
#> iter  30 value 357.894435
#> iter  40 value 345.126177
#> iter  50 value 344.871878
#> final  value 344.869464 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 488.746061
#> iter  20 value 377.661320
#> iter  30 value 356.923705
#> iter  40 value 349.034794
#> iter  50 value 348.951754
#> final  value 348.951618 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 476.077154
#> iter  20 value 371.376069
#> iter  30 value 357.895907
#> iter  40 value 345.130291
#> iter  50 value 344.876338
#> final  value 344.873916 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 747.212661 
#> iter  10 value 468.672664
#> iter  20 value 377.451939
#> iter  30 value 358.210489
#> iter  40 value 349.724359
#> iter  50 value 349.580936
#> final  value 349.580795 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 747.212661 
#> iter  10 value 479.627477
#> iter  20 value 383.927129
#> iter  30 value 358.552621
#> iter  40 value 353.786394
#> iter  50 value 353.658844
#> final  value 353.658312 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 747.212661 
#> iter  10 value 468.684467
#> iter  20 value 377.463622
#> iter  30 value 358.217236
#> iter  40 value 349.729193
#> iter  50 value 349.585408
#> final  value 349.585268 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 480.520903
#> iter  20 value 377.444154
#> iter  30 value 353.128165
#> iter  40 value 345.608787
#> iter  50 value 345.459556
#> final  value 345.459445 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 493.821669
#> iter  20 value 380.135940
#> iter  30 value 353.605207
#> iter  40 value 349.572271
#> iter  50 value 349.533239
#> final  value 349.533226 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 480.535164
#> iter  20 value 377.455658
#> iter  30 value 353.134404
#> iter  40 value 345.613835
#> iter  50 value 345.463984
#> final  value 345.463874 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 752.757838 
#> iter  10 value 367.531721
#> iter  20 value 349.997188
#> iter  30 value 348.181552
#> iter  40 value 347.970064
#> iter  50 value 347.968093
#> final  value 347.968087 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 752.757838 
#> iter  10 value 369.824154
#> iter  20 value 353.578640
#> iter  30 value 352.143721
#> iter  40 value 352.025807
#> iter  50 value 352.023970
#> final  value 352.023965 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 752.757838 
#> iter  10 value 367.533996
#> iter  20 value 350.000476
#> iter  30 value 348.185944
#> iter  40 value 347.974513
#> iter  50 value 347.972542
#> final  value 347.972535 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 482.372644
#> iter  20 value 382.561764
#> iter  30 value 365.586309
#> iter  40 value 351.133992
#> iter  50 value 350.184871
#> final  value 350.143464 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 493.340773
#> iter  20 value 386.192398
#> iter  30 value 362.534927
#> iter  40 value 354.321035
#> iter  50 value 354.158400
#> final  value 354.156458 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 482.384631
#> iter  20 value 382.572995
#> iter  30 value 365.592078
#> iter  40 value 351.082152
#> iter  50 value 350.188873
#> final  value 350.147823 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 502.564298
#> iter  20 value 379.743268
#> iter  30 value 353.891103
#> iter  40 value 346.896577
#> iter  50 value 346.664529
#> final  value 346.661877 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 513.846068
#> iter  20 value 386.210499
#> iter  30 value 357.531882
#> iter  40 value 351.135033
#> iter  50 value 350.872906
#> final  value 350.871442 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 502.576719
#> iter  20 value 379.751023
#> iter  30 value 353.892884
#> iter  40 value 346.902228
#> iter  50 value 346.669800
#> final  value 346.666513 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  56 (39 variable)
#> initial  value 833.162911 
#> iter  10 value 408.578154
#> iter  20 value 393.672258
#> iter  30 value 392.470009
#> iter  40 value 392.328865
#> final  value 392.326118 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex      married        age    havejob        educ
#> Formal     -2.744085 -0.08052509  0.036896047 -1.0830442 0.31369245  0.27880615
#> Informal   -2.818296 -0.11941548 -0.005803761 -0.5875136 0.07289354 -0.00319335
#> L.Both     -3.299207  0.02057550 -0.270602324 -0.9213954 0.49270329  0.14584711
#>          political.afl      rural      region fin.intermdiaries fin.knowldge
#> Formal       0.2687575  0.3512704 -0.21502157        0.08159645   0.17375188
#> Informal    -0.4672335 -0.1134654  0.25575221       -0.14683815   0.12374445
#> L.Both       0.2390123 -0.2293460  0.06995488        0.21848184   0.04403905
#>              income   networth
#> Formal    0.7099168  0.4001333
#> Informal -2.3272713 -1.0252919
#> L.Both    0.3565595  0.1332427
#> 
#> Residual Deviance: 784.6522 
#> AIC: 862.6522 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7215