Random Forest

Random Forest

Usage

RF_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from RF_Model.

Details

Rather than considering the random sample of \(m\) predictors from the total of \(p\) predictors in each split, random forest does not consider a majority of the \(p\) predictors, and considers in each split a fresh sample of \(m_{try}\) which we usually set to \(m_{try} \approx \sqrt{p}\) Random forests which de-correlate the trees by considering \(m_{try} \approx \sqrt{p}\) show an improvement over bagged trees \(m = p\).

Examples

# \donttest{
yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:250),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLR.1 <- RF_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold01: mtry= 2 
#> + Fold01: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold01: mtry= 6 
#> + Fold01: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold01: mtry=10 
#> + Fold02: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold02: mtry= 2 
#> + Fold02: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold02: mtry= 6 
#> + Fold02: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold02: mtry=10 
#> + Fold03: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold03: mtry= 2 
#> + Fold03: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold03: mtry= 6 
#> + Fold03: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold03: mtry=10 
#> + Fold04: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold04: mtry= 2 
#> + Fold04: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold04: mtry= 6 
#> + Fold04: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold04: mtry=10 
#> + Fold05: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold05: mtry= 2 
#> + Fold05: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold05: mtry= 6 
#> + Fold05: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold05: mtry=10 
#> + Fold06: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold06: mtry= 2 
#> + Fold06: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold06: mtry= 6 
#> + Fold06: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold06: mtry=10 
#> + Fold07: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold07: mtry= 2 
#> + Fold07: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold07: mtry= 6 
#> + Fold07: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold07: mtry=10 
#> + Fold08: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold08: mtry= 2 
#> + Fold08: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold08: mtry= 6 
#> + Fold08: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold08: mtry=10 
#> + Fold09: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold09: mtry= 2 
#> + Fold09: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold09: mtry= 6 
#> + Fold09: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold09: mtry=10 
#> + Fold10: mtry= 2 
#> Warning: These variables have zero variances: region
#> - Fold10: mtry= 2 
#> + Fold10: mtry= 6 
#> Warning: These variables have zero variances: region
#> - Fold10: mtry= 6 
#> + Fold10: mtry=10 
#> Warning: These variables have zero variances: region
#> - Fold10: mtry=10 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting mtry = 2 on full training set
#> Warning: These variables have zero variances: region
BchMk.MLR.1$finalModel
#> 
#> Call:
#>  randomForest(formula = Data.sub.train[, yvar] ~ ., data = Data.sub.train) 
#>                Type of random forest: classification
#>                      Number of trees: 500
#> No. of variables tried at each split: 2
#> 
#>         OOB estimate of  error rate: 27.36%
#> Confusion matrix:
#>          No.Loan Formal Informal L.Both class.error
#> No.Loan      123     10        1      0  0.08208955
#> Formal        18     23        0      0  0.43902439
#> Informal      16      0        0      0  1.00000000
#> L.Both         5      4        1      0  1.00000000
BchMk.MLR.1$Roc$auc
#> Multi-class area under the curve: 0.8011
 # }