Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 367.671793
#> iter  20 value 348.920917
#> iter  30 value 347.263782
#> iter  40 value 347.186337
#> final  value 347.186228 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 369.804490
#> iter  20 value 352.313903
#> iter  30 value 350.886495
#> iter  40 value 350.854883
#> final  value 350.854758 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 367.673945
#> iter  20 value 348.924402
#> iter  30 value 347.252729
#> iter  40 value 347.190169
#> final  value 347.190046 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 363.696340
#> iter  20 value 348.924470
#> iter  30 value 347.731159
#> iter  40 value 347.680338
#> final  value 347.680144 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 365.907634
#> iter  20 value 352.159445
#> iter  30 value 351.188668
#> iter  40 value 351.151900
#> final  value 351.151807 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 363.698570
#> iter  20 value 348.927786
#> iter  30 value 347.734752
#> iter  40 value 347.683931
#> final  value 347.683737 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 373.589951
#> iter  20 value 356.250776
#> iter  30 value 355.161062
#> iter  40 value 355.135854
#> final  value 355.135450 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 375.673531
#> iter  20 value 359.422160
#> iter  30 value 358.419611
#> iter  40 value 358.387461
#> final  value 358.387369 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 373.592053
#> iter  20 value 356.254015
#> iter  30 value 355.164408
#> iter  40 value 355.139201
#> final  value 355.138796 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 372.296688
#> iter  20 value 355.807428
#> iter  30 value 354.286583
#> iter  40 value 354.247494
#> final  value 354.247384 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 374.466919
#> iter  20 value 358.399114
#> iter  30 value 357.613802
#> iter  40 value 357.589300
#> final  value 357.589194 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 372.298880
#> iter  20 value 355.811063
#> iter  30 value 354.290023
#> iter  40 value 354.250942
#> final  value 354.250832 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 361.202726
#> iter  20 value 344.211067
#> iter  30 value 342.671945
#> iter  40 value 342.632588
#> final  value 342.632412 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 363.261140
#> iter  20 value 347.373868
#> iter  30 value 346.123232
#> iter  40 value 346.095146
#> final  value 346.095045 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 361.204800
#> iter  20 value 344.214297
#> iter  30 value 342.675504
#> iter  40 value 342.636176
#> final  value 342.635999 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 365.599930
#> iter  20 value 348.684575
#> iter  30 value 347.115864
#> iter  40 value 347.072715
#> final  value 347.072657 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 367.674324
#> iter  20 value 351.839434
#> iter  30 value 350.614996
#> iter  40 value 350.539753
#> final  value 350.539478 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 365.602020
#> iter  20 value 348.687794
#> iter  30 value 347.119461
#> iter  40 value 347.076306
#> final  value 347.076248 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 369.076386
#> iter  20 value 351.181582
#> iter  30 value 349.839792
#> iter  40 value 349.803484
#> final  value 349.803331 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 371.137628
#> iter  20 value 354.298615
#> iter  30 value 353.223507
#> iter  40 value 353.198954
#> final  value 353.198915 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 369.078465
#> iter  20 value 351.184768
#> iter  30 value 349.843287
#> iter  40 value 349.807000
#> final  value 349.806847 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 365.970058
#> iter  20 value 348.270993
#> iter  30 value 347.037431
#> iter  40 value 346.980488
#> final  value 346.980360 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 368.065999
#> iter  20 value 351.474477
#> iter  30 value 350.441538
#> iter  40 value 350.416401
#> final  value 350.416330 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 365.972172
#> iter  20 value 348.274258
#> iter  30 value 347.041029
#> iter  40 value 346.984036
#> final  value 346.983908 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 366.230880
#> iter  20 value 349.091272
#> iter  30 value 347.248005
#> iter  40 value 347.181389
#> final  value 347.181281 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 368.303620
#> iter  20 value 353.212869
#> iter  30 value 350.794425
#> iter  40 value 350.721609
#> final  value 350.721288 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 366.232969
#> iter  20 value 349.094903
#> iter  30 value 347.251657
#> iter  40 value 347.185058
#> final  value 347.184949 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 371.201596
#> iter  20 value 354.332341
#> iter  30 value 353.344233
#> final  value 353.318036 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 373.227554
#> iter  20 value 357.404619
#> iter  30 value 356.633548
#> iter  40 value 356.598124
#> final  value 356.598024 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 371.203638
#> iter  20 value 354.335478
#> iter  30 value 353.347630
#> final  value 353.321418 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  48 (33 variable)
#> initial  value 833.162911 
#> iter  10 value 409.146026
#> iter  20 value 393.785618
#> iter  30 value 393.079597
#> iter  40 value 393.032525
#> final  value 393.032437 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex     married        age    havejob        educ
#> Formal     -2.833495  0.05443850  0.41408093 -1.0558778  0.1697851  0.36553102
#> Informal   -2.711203 -0.08039109 -0.09041137 -0.7971489 -0.2208282 -0.07134397
#> L.Both     -3.370696  0.05283664  0.06185204 -0.8789984  0.4253913  0.47669894
#>               rural      region     income political.afl   networth
#> Formal    0.4858247 -0.49700532  0.2977503   0.137964524  0.3013226
#> Informal -0.2271720  0.27110732 -0.7327700  -0.466147896 -1.0251643
#> L.Both   -0.1327603 -0.04648268  0.1211763   0.008906008  0.1573417
#> 
#> Residual Deviance: 786.0649 
#> AIC: 852.0649 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.6933