Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
xvar <- c("sex", "married", "age", "havejob", "educ", "political.afl",
"rural", "region", "fin.intermdiaries", "fin.knowldge", "income")
BchMk.MLM <- MLM_Model(sample_data, c(xvar, "networth"), yvar )
#> + Fold01: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 373.723158
#> iter  20 value 360.721473
#> iter  30 value 359.140224
#> iter  40 value 359.042439
#> final  value 359.041346 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 375.734327
#> iter  20 value 363.649996
#> iter  30 value 362.456539
#> iter  40 value 362.388528
#> final  value 362.387968 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 373.725187
#> iter  20 value 360.724476
#> iter  30 value 359.143715
#> iter  40 value 359.045993
#> final  value 359.044897 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 368.242504
#> iter  20 value 352.559277
#> iter  30 value 350.660359
#> iter  40 value 350.369318
#> iter  50 value 350.343544
#> final  value 350.341695 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 370.221623
#> iter  20 value 355.734122
#> iter  30 value 353.877973
#> iter  40 value 353.714365
#> iter  50 value 353.711390
#> iter  50 value 353.711390
#> iter  50 value 353.711390
#> final  value 353.711390 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 368.244500
#> iter  20 value 352.562437
#> iter  30 value 350.663410
#> iter  40 value 350.372817
#> iter  50 value 350.347272
#> final  value 350.345745 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 752.757838 
#> iter  10 value 381.950799
#> iter  20 value 370.310632
#> iter  30 value 369.503730
#> iter  40 value 369.425715
#> final  value 369.425347 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 752.757838 
#> iter  10 value 383.933671
#> iter  20 value 373.055450
#> iter  30 value 372.420826
#> iter  40 value 372.359625
#> final  value 372.359494 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 752.757838 
#> iter  10 value 381.952799
#> iter  20 value 370.313435
#> iter  30 value 369.506725
#> iter  40 value 369.428744
#> final  value 369.428376 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 373.893290
#> iter  20 value 361.820305
#> iter  30 value 360.850162
#> iter  40 value 360.752609
#> iter  50 value 360.727062
#> final  value 360.726022 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 375.933962
#> iter  20 value 364.686856
#> iter  30 value 363.892746
#> iter  40 value 363.843442
#> iter  50 value 363.837808
#> iter  50 value 363.837805
#> iter  50 value 363.837805
#> final  value 363.837805 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 373.895349
#> iter  20 value 361.823220
#> iter  30 value 360.853307
#> iter  40 value 360.755852
#> iter  50 value 360.730595
#> final  value 360.729869 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 376.332927
#> iter  20 value 364.072840
#> iter  30 value 362.984085
#> iter  40 value 362.903843
#> final  value 362.903254 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 378.296707
#> iter  20 value 366.923797
#> iter  30 value 365.938396
#> iter  40 value 365.860821
#> final  value 365.858337 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 376.334907
#> iter  20 value 364.075733
#> iter  30 value 362.987115
#> iter  40 value 362.906884
#> final  value 362.906296 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 747.212661 
#> iter  10 value 370.827283
#> iter  20 value 357.075376
#> iter  30 value 355.789803
#> iter  40 value 355.644708
#> iter  50 value 355.642926
#> final  value 355.642906 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 747.212661 
#> iter  10 value 372.788245
#> iter  20 value 360.031131
#> iter  30 value 358.830515
#> iter  40 value 358.731016
#> final  value 358.730631 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 747.212661 
#> iter  10 value 370.829260
#> iter  20 value 357.078393
#> iter  30 value 355.792862
#> iter  40 value 355.647899
#> iter  50 value 355.646120
#> final  value 355.646099 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 377.881612
#> iter  20 value 363.012347
#> iter  30 value 362.194450
#> iter  40 value 362.096198
#> final  value 362.095723 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 379.806483
#> iter  20 value 365.861874
#> iter  30 value 365.164287
#> iter  40 value 365.122827
#> final  value 365.122707 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 377.883552
#> iter  20 value 363.015254
#> iter  30 value 362.197587
#> iter  40 value 362.099326
#> final  value 362.098851 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 375.858335
#> iter  20 value 362.003505
#> iter  30 value 360.978936
#> iter  40 value 360.903334
#> final  value 360.902719 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 377.846238
#> iter  20 value 364.868179
#> iter  30 value 364.021352
#> iter  40 value 363.955168
#> final  value 363.954467 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 375.860340
#> iter  20 value 362.006423
#> iter  30 value 360.982065
#> iter  40 value 360.906482
#> final  value 360.905868 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 372.331634
#> iter  20 value 358.214120
#> iter  30 value 357.107147
#> iter  40 value 356.998513
#> final  value 356.997053 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 374.302089
#> iter  20 value 361.055485
#> iter  30 value 360.112515
#> iter  40 value 360.033795
#> final  value 360.032532 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 372.333620
#> iter  20 value 358.217016
#> iter  30 value 357.110234
#> iter  40 value 357.001638
#> final  value 357.000176 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 374.974546
#> iter  20 value 362.771295
#> iter  30 value 361.864675
#> iter  40 value 361.786068
#> final  value 361.785621 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 376.939918
#> iter  20 value 365.609070
#> iter  30 value 364.800991
#> iter  40 value 364.762828
#> final  value 364.762671 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 374.976528
#> iter  20 value 362.774171
#> iter  30 value 361.867740
#> iter  40 value 361.789145
#> final  value 361.788698 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  56 (39 variable)
#> initial  value 833.162911 
#> iter  10 value 417.350810
#> iter  20 value 406.132907
#> iter  30 value 405.402205
#> iter  40 value 405.376513
#> final  value 405.376379 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)          sex    married        age     havejob        educ
#> Formal     -2.541666 -0.002656331  0.1480069 -0.8050364  0.26045310  0.07302542
#> Informal   -2.635864 -0.047332030 -0.1431530 -0.7291428 -0.12335096 -0.12546252
#> L.Both     -3.244954  0.007497894 -0.2150453 -0.9904541  0.04430019  0.29514385
#>          political.afl      rural      region fin.intermdiaries fin.knowldge
#> Formal      0.20322402  0.3890704 -0.25339785       -0.03200172   0.12662317
#> Informal   -0.27268885 -0.1386307  0.21184726       -0.21106483  -0.02069926
#> L.Both     -0.03896539 -0.1852625 -0.05609517        0.05273786  -0.03518182
#>              income    networth
#> Formal    0.4384895  0.39498688
#> Informal -0.5841380 -1.18708720
#> L.Both    0.2896624  0.04193885
#> 
#> Residual Deviance: 810.7528 
#> AIC: 888.7528 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7849