Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 443.934581
#> iter  20 value 367.455647
#> iter  30 value 344.241732
#> iter  40 value 341.437625
#> final  value 341.433950 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 453.899966
#> iter  20 value 376.038933
#> iter  30 value 348.741232
#> iter  40 value 345.774261
#> final  value 345.767345 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 443.945372
#> iter  20 value 367.462765
#> iter  30 value 344.182485
#> iter  40 value 341.441504
#> final  value 341.438661 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 464.062077
#> iter  20 value 370.540949
#> iter  30 value 348.318452
#> iter  40 value 346.317090
#> final  value 346.315039 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 403.077922
#> iter  20 value 369.253148
#> iter  30 value 350.738112
#> iter  40 value 350.238026
#> final  value 350.237692 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 464.070597
#> iter  20 value 370.546467
#> iter  30 value 348.322871
#> iter  40 value 346.321297
#> final  value 346.319260 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 446.752464
#> iter  20 value 369.427130
#> iter  30 value 345.067260
#> iter  40 value 343.715573
#> final  value 343.714657 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 454.936715
#> iter  20 value 373.004392
#> iter  30 value 348.969893
#> iter  40 value 347.530661
#> final  value 347.526516 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 446.761596
#> iter  20 value 369.429058
#> iter  30 value 345.062779
#> iter  40 value 343.719775
#> final  value 343.718725 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 442.701851
#> iter  20 value 363.861894
#> iter  30 value 344.501172
#> iter  40 value 343.321033
#> final  value 343.318465 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 450.856661
#> iter  20 value 369.113671
#> iter  30 value 348.092067
#> iter  40 value 347.317101
#> final  value 347.316917 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 442.710918
#> iter  20 value 363.866189
#> iter  30 value 344.503512
#> iter  40 value 343.325368
#> final  value 343.322787 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 450.124520
#> iter  20 value 368.994462
#> iter  30 value 344.031819
#> iter  40 value 342.634960
#> final  value 342.629815 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 458.601858
#> iter  20 value 370.602378
#> iter  30 value 348.078959
#> iter  40 value 346.570154
#> final  value 346.568085 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 450.133833
#> iter  20 value 368.962459
#> iter  30 value 343.993038
#> iter  40 value 342.639870
#> final  value 342.634039 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 424.140319
#> iter  20 value 353.011097
#> iter  30 value 337.404270
#> iter  40 value 333.191904
#> final  value 333.163058 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 434.631148
#> iter  20 value 360.064301
#> iter  30 value 338.597818
#> iter  40 value 337.576074
#> final  value 337.575796 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 424.151670
#> iter  20 value 353.020552
#> iter  30 value 337.386107
#> iter  40 value 333.178211
#> final  value 333.167862 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 438.114116
#> iter  20 value 364.625123
#> iter  30 value 341.443944
#> iter  40 value 340.358249
#> final  value 340.353457 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 447.289739
#> iter  20 value 365.901827
#> iter  30 value 345.567378
#> iter  40 value 344.583411
#> final  value 344.581661 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 438.124134
#> iter  20 value 364.631746
#> iter  30 value 341.445479
#> iter  40 value 340.363005
#> final  value 340.358027 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 480.508429
#> iter  20 value 385.191129
#> iter  30 value 349.283901
#> iter  40 value 347.096559
#> final  value 347.094642 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 491.302440
#> iter  20 value 386.357322
#> iter  30 value 351.582597
#> iter  40 value 350.642927
#> final  value 350.641690 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 480.520191
#> iter  20 value 385.199544
#> iter  30 value 349.290822
#> iter  40 value 347.100272
#> final  value 347.098311 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 446.055223
#> iter  20 value 364.101143
#> iter  30 value 337.920750
#> iter  40 value 335.566724
#> final  value 335.563537 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 459.487351
#> iter  20 value 365.580355
#> iter  30 value 341.143142
#> iter  40 value 339.532806
#> final  value 339.531491 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 446.069562
#> iter  20 value 364.101394
#> iter  30 value 337.927430
#> iter  40 value 335.570810
#> final  value 335.567742 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 440.980472
#> iter  20 value 363.066646
#> iter  30 value 347.269146
#> iter  40 value 341.881085
#> final  value 341.873723 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 451.447064
#> iter  20 value 372.909614
#> iter  30 value 347.739158
#> iter  40 value 345.932404
#> final  value 345.931622 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 440.992328
#> iter  20 value 363.081195
#> iter  30 value 347.236623
#> iter  40 value 341.879301
#> final  value 341.878033 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  48 (33 variable)
#> initial  value 833.162911 
#> iter  10 value 547.039457
#> iter  20 value 429.959997
#> iter  30 value 388.869335
#> iter  40 value 385.344892
#> final  value 385.339029 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)          sex    married        age     havejob        educ
#> Formal     -2.985010 -0.096030402  0.3445827 -1.1972301  0.12272190  0.37522893
#> Informal   -2.703773 -0.002796409 -0.1772952 -0.6940295 -0.05452653 -0.07854443
#> L.Both     -3.340874  0.030399343 -0.2076060 -1.0573222  0.13597561  0.51164632
#>               rural      region     income political.afl   networth
#> Formal    0.5688181 -0.44416916  0.7403666    0.16083928  0.5652596
#> Informal -0.1709390  0.21368389 -1.7770967   -0.16894533 -1.2125471
#> L.Both   -0.1180169  0.08001169 -0.3025295    0.02176402  0.1807334
#> 
#> Residual Deviance: 770.6781 
#> AIC: 836.6781 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.676