Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
xvar <- c("sex", "married", "age", "havejob", "educ", "political.afl",
"rural", "region", "fin.intermdiaries", "fin.knowldge", "income")
BchMk.MLM <- MLM_Model(sample_data, c(xvar, "networth"), yvar )
#> + Fold01: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 747.212661 
#> iter  10 value 355.496032
#> iter  20 value 339.613215
#> iter  30 value 337.967317
#> iter  40 value 337.797070
#> final  value 337.795335 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 747.212661 
#> iter  10 value 357.696885
#> iter  20 value 343.210840
#> iter  30 value 341.996702
#> iter  40 value 341.847285
#> iter  50 value 341.844036
#> final  value 341.844012 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 747.212661 
#> iter  10 value 355.498247
#> iter  20 value 339.616990
#> iter  30 value 337.971678
#> iter  40 value 337.801473
#> final  value 337.799737 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 417.518520
#> iter  20 value 376.952102
#> iter  30 value 351.706978
#> iter  40 value 349.057121
#> iter  50 value 348.948026
#> final  value 348.948002 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 366.832673
#> iter  20 value 354.521915
#> iter  30 value 352.983469
#> iter  40 value 352.857252
#> iter  50 value 352.850775
#> iter  50 value 352.850772
#> iter  50 value 352.850772
#> final  value 352.850772 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 417.532013
#> iter  20 value 376.950694
#> iter  30 value 351.709499
#> iter  40 value 349.060562
#> iter  50 value 348.952261
#> final  value 348.952236 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 362.319094
#> iter  20 value 344.845393
#> iter  30 value 342.821646
#> iter  40 value 342.639973
#> final  value 342.632877 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 364.409913
#> iter  20 value 348.304118
#> iter  30 value 346.887728
#> iter  40 value 346.750532
#> iter  50 value 346.740183
#> iter  50 value 346.740182
#> iter  50 value 346.740182
#> final  value 346.740182 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 362.321206
#> iter  20 value 344.849051
#> iter  30 value 342.826039
#> iter  40 value 342.644448
#> final  value 342.637349 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 462.756221
#> iter  20 value 373.561934
#> iter  30 value 342.378726
#> iter  40 value 334.979352
#> iter  50 value 334.914409
#> final  value 334.914288 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 476.367349
#> iter  20 value 376.051204
#> iter  30 value 344.842908
#> iter  40 value 339.374952
#> iter  50 value 339.306809
#> final  value 339.306786 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 462.770937
#> iter  20 value 373.575907
#> iter  30 value 342.339830
#> iter  40 value 334.984587
#> iter  50 value 334.919187
#> final  value 334.919062 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 487.211558
#> iter  20 value 378.261650
#> iter  30 value 350.969801
#> iter  40 value 343.871322
#> iter  50 value 343.447192
#> final  value 343.447154 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 497.504123
#> iter  20 value 383.255371
#> iter  30 value 352.369636
#> iter  40 value 347.408977
#> iter  50 value 347.338809
#> iter  50 value 347.338807
#> iter  50 value 347.338807
#> final  value 347.338807 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 487.223021
#> iter  20 value 378.272013
#> iter  30 value 350.951061
#> iter  40 value 343.876589
#> iter  50 value 343.451356
#> final  value 343.451317 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 467.561492
#> iter  20 value 369.283279
#> iter  30 value 341.696072
#> iter  40 value 337.511894
#> iter  50 value 337.461298
#> final  value 337.461277 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 480.540460
#> iter  20 value 374.603883
#> iter  30 value 350.213957
#> iter  40 value 342.232151
#> iter  50 value 341.900334
#> final  value 341.900165 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 467.575634
#> iter  20 value 369.294791
#> iter  30 value 341.722677
#> iter  40 value 337.533527
#> iter  50 value 337.482350
#> iter  60 value 337.470402
#> final  value 337.466848 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 481.200603
#> iter  20 value 368.663456
#> iter  30 value 345.536684
#> iter  40 value 340.105230
#> iter  50 value 340.037450
#> final  value 340.037375 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 494.609715
#> iter  20 value 374.789620
#> iter  30 value 351.929506
#> iter  40 value 344.177161
#> iter  50 value 344.028929
#> final  value 344.028902 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 481.215214
#> iter  20 value 368.673647
#> iter  30 value 345.558907
#> iter  40 value 340.110093
#> iter  50 value 340.041742
#> final  value 340.041720 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 752.757838 
#> iter  10 value 358.740263
#> iter  20 value 346.550070
#> iter  30 value 344.382799
#> iter  40 value 344.155920
#> iter  50 value 344.149130
#> iter  50 value 344.149129
#> iter  50 value 344.149129
#> final  value 344.149129 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 752.757838 
#> iter  10 value 360.928637
#> iter  20 value 349.344713
#> iter  30 value 347.804810
#> iter  40 value 347.635450
#> iter  50 value 347.630216
#> iter  50 value 347.630216
#> iter  50 value 347.630216
#> final  value 347.630216 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 752.757838 
#> iter  10 value 358.742470
#> iter  20 value 346.553497
#> iter  30 value 344.386457
#> iter  40 value 344.159555
#> iter  50 value 344.152755
#> iter  50 value 344.152755
#> iter  50 value 344.152755
#> final  value 344.152755 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 472.262275
#> iter  20 value 371.930442
#> iter  30 value 341.148306
#> iter  40 value 335.761207
#> iter  50 value 335.670455
#> final  value 335.670446 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 360.899229
#> iter  20 value 341.804157
#> iter  30 value 340.240964
#> iter  40 value 340.122386
#> iter  50 value 340.118313
#> iter  50 value 340.118311
#> iter  50 value 340.118311
#> final  value 340.118311 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 472.277293
#> iter  20 value 371.951047
#> iter  30 value 341.156224
#> iter  40 value 335.768688
#> iter  50 value 335.675295
#> final  value 335.675280 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 488.097010
#> iter  20 value 378.330792
#> iter  30 value 351.118983
#> iter  40 value 344.834101
#> iter  50 value 344.751577
#> final  value 344.751437 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 499.851997
#> iter  20 value 383.277110
#> iter  30 value 352.077691
#> iter  40 value 348.632304
#> iter  50 value 348.463134
#> final  value 348.463114 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 488.109917
#> iter  20 value 378.337432
#> iter  30 value 351.118071
#> iter  40 value 344.836273
#> iter  50 value 344.755478
#> final  value 344.755352 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  56 (39 variable)
#> initial  value 833.162911 
#> iter  10 value 400.774042
#> iter  20 value 386.937070
#> iter  30 value 385.307210
#> iter  40 value 385.060857
#> iter  50 value 385.060088
#> final  value 385.060079 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex    married        age     havejob       educ
#> Formal     -2.847544 -0.12756702  0.3490328 -1.0638978  0.23527310  0.1967655
#> Informal   -2.837989  0.06153095 -0.2372119 -0.6584559 -0.03956239 -0.1418404
#> L.Both     -3.205824 -0.08053778 -0.2069994 -0.6602548  0.26413302  0.5030912
#>          political.afl      rural     region fin.intermdiaries fin.knowldge
#> Formal      0.31161423  0.4755303 -0.4310751       0.001617175   0.15224052
#> Informal   -0.27658573 -0.2892537  0.2687106      -0.224904696   0.17816452
#> L.Both     -0.05807255 -0.2930099 -0.1889835       0.189691792  -0.04275343
#>             income    networth
#> Formal    1.073501  0.36935690
#> Informal -1.957277 -1.09713449
#> L.Both    0.924945  0.04671119
#> 
#> Residual Deviance: 770.1202 
#> AIC: 848.1202 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.6935