Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
xvar <- c("sex", "married", "age", "havejob", "educ", "political.afl",
"rural", "region", "fin.intermdiaries", "fin.knowldge", "income")
BchMk.MLM <- MLM_Model(sample_data, c(xvar, "networth"), yvar )
#> + Fold01: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 506.209628
#> iter  20 value 384.236762
#> iter  30 value 360.833421
#> iter  40 value 355.450144
#> iter  50 value 355.287585
#> final  value 355.287545 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 423.839563
#> iter  20 value 386.597518
#> iter  30 value 361.164027
#> iter  40 value 358.967298
#> iter  50 value 358.913005
#> iter  50 value 358.913003
#> iter  50 value 358.913003
#> final  value 358.913003 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 506.221433
#> iter  20 value 384.250118
#> iter  30 value 360.830019
#> iter  40 value 355.452252
#> iter  50 value 355.291436
#> final  value 355.291396 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 498.747959
#> iter  20 value 379.009462
#> iter  30 value 357.084414
#> iter  40 value 349.201025
#> iter  50 value 349.101543
#> final  value 349.101388 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 412.721055
#> iter  20 value 381.893023
#> iter  30 value 354.877541
#> iter  40 value 352.900082
#> iter  50 value 352.805489
#> final  value 352.805135 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 498.757182
#> iter  20 value 379.018312
#> iter  30 value 357.085160
#> iter  40 value 349.205828
#> iter  50 value 349.105470
#> final  value 349.105316 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 498.409018
#> iter  20 value 383.451954
#> iter  30 value 357.141126
#> iter  40 value 352.817845
#> iter  50 value 352.691687
#> final  value 352.691659 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 421.405176
#> iter  20 value 379.169416
#> iter  30 value 359.467486
#> iter  40 value 356.364762
#> iter  50 value 356.333645
#> final  value 356.333626 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 498.418913
#> iter  20 value 383.438290
#> iter  30 value 357.146499
#> iter  40 value 352.828589
#> iter  50 value 352.695620
#> final  value 352.695541 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 518.981204
#> iter  20 value 390.214534
#> iter  30 value 361.720979
#> iter  40 value 357.293998
#> iter  50 value 357.054862
#> final  value 357.052165 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 528.822078
#> iter  20 value 398.135469
#> iter  30 value 365.460839
#> iter  40 value 360.614925
#> iter  50 value 360.495299
#> final  value 360.494767 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 518.993295
#> iter  20 value 390.221812
#> iter  30 value 361.728448
#> iter  40 value 357.298728
#> iter  50 value 357.058474
#> final  value 357.055812 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 422.878970
#> iter  20 value 384.295152
#> iter  30 value 360.292707
#> iter  40 value 354.408399
#> iter  50 value 354.240640
#> final  value 354.240611 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 377.255718
#> iter  20 value 359.094104
#> iter  30 value 358.090313
#> iter  40 value 358.036940
#> iter  50 value 358.036468
#> final  value 358.036420 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 422.892258
#> iter  20 value 384.299811
#> iter  30 value 360.308523
#> iter  40 value 354.415214
#> iter  50 value 354.244765
#> final  value 354.244748 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 489.664491
#> iter  20 value 391.701341
#> iter  30 value 357.648510
#> iter  40 value 349.322942
#> iter  50 value 349.123978
#> final  value 349.123955 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 501.909649
#> iter  20 value 395.018203
#> iter  30 value 359.433062
#> iter  40 value 353.221205
#> iter  50 value 352.997757
#> final  value 352.997690 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 489.677751
#> iter  20 value 391.700725
#> iter  30 value 357.639310
#> iter  40 value 349.338781
#> iter  50 value 349.128140
#> final  value 349.128111 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 483.532925
#> iter  20 value 388.721848
#> iter  30 value 365.509561
#> iter  40 value 355.510824
#> iter  50 value 354.609250
#> final  value 354.607977 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 492.222394
#> iter  20 value 395.522132
#> iter  30 value 362.437336
#> iter  40 value 358.054768
#> iter  50 value 358.018657
#> iter  50 value 358.018654
#> iter  50 value 358.018654
#> final  value 358.018654 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 483.542583
#> iter  20 value 388.734866
#> iter  30 value 365.511493
#> iter  40 value 355.537534
#> iter  50 value 354.612885
#> final  value 354.611572 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 366.582717
#> iter  20 value 350.036047
#> iter  30 value 348.868897
#> iter  40 value 348.761028
#> iter  50 value 348.759044
#> iter  50 value 348.759043
#> iter  50 value 348.759043
#> final  value 348.759043 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 368.589255
#> iter  20 value 353.334020
#> iter  30 value 352.416076
#> iter  40 value 352.362775
#> final  value 352.361244 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 366.584744
#> iter  20 value 350.039458
#> iter  30 value 348.872670
#> iter  40 value 348.764828
#> iter  50 value 348.762843
#> iter  50 value 348.762842
#> iter  50 value 348.762842
#> final  value 348.762842 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 409.398734
#> iter  20 value 375.297070
#> iter  30 value 359.304512
#> iter  40 value 356.792742
#> iter  50 value 356.757998
#> iter  50 value 356.757997
#> iter  50 value 356.757997
#> final  value 356.757997 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 421.580286
#> iter  20 value 377.497320
#> iter  30 value 365.023804
#> iter  40 value 360.630833
#> iter  50 value 360.344404
#> final  value 360.344347 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 409.411216
#> iter  20 value 375.305341
#> iter  30 value 359.322243
#> iter  40 value 356.810266
#> iter  50 value 356.775496
#> iter  60 value 356.762665
#> final  value 356.762306 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 367.735182
#> iter  20 value 353.413040
#> iter  30 value 352.265766
#> iter  40 value 352.091900
#> final  value 352.089949 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 369.785750
#> iter  20 value 356.404982
#> iter  30 value 355.479847
#> iter  40 value 355.357065
#> final  value 355.355443 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 367.737250
#> iter  20 value 353.416103
#> iter  30 value 352.269077
#> iter  40 value 352.095282
#> final  value 352.093331 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 1e-04 on full training set
#> # weights:  56 (39 variable)
#> initial  value 833.162911 
#> iter  10 value 412.580108
#> iter  20 value 395.484826
#> iter  30 value 394.474574
#> iter  40 value 394.410878
#> final  value 394.410325 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex    married        age    havejob        educ
#> Formal     -2.698662 -0.21865806  0.2956626 -1.0447791 0.21556715  0.18303709
#> Informal   -2.788642 -0.17594671 -0.1016013 -0.6373116 0.07139939 -0.05253701
#> L.Both     -3.406405  0.01156829 -0.2244586 -1.1008824 0.18481408  0.30394865
#>          political.afl      rural       region fin.intermdiaries fin.knowldge
#> Formal      0.25801825  0.3551217 -0.246815286        0.04362295   0.23032226
#> Informal   -0.34335288 -0.1917323  0.105678333       -0.13223849   0.16049012
#> L.Both      0.07191478 -0.4018127 -0.008530078        0.23821083   0.04082711
#>              income   networth
#> Formal    0.7105534  0.3769751
#> Informal -1.9799224 -1.2142562
#> L.Both    0.4025622  0.2727291
#> 
#> Residual Deviance: 788.8207 
#> AIC: 866.8207 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7342