Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 483.860705
#> iter  20 value 386.087905
#> iter  30 value 358.940487
#> iter  40 value 355.950018
#> final  value 355.940818 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 489.388716
#> iter  20 value 388.906834
#> iter  30 value 360.460804
#> iter  40 value 359.165915
#> final  value 359.158075 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 483.866977
#> iter  20 value 386.097342
#> iter  30 value 358.941165
#> iter  40 value 355.953259
#> final  value 355.944154 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 374.216303
#> iter  20 value 361.795947
#> iter  30 value 360.504079
#> iter  40 value 360.385654
#> final  value 360.384958 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 376.179825
#> iter  20 value 364.659978
#> iter  30 value 363.480728
#> iter  40 value 363.375713
#> iter  40 value 363.375711
#> iter  40 value 363.375709
#> final  value 363.375709 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 374.218281
#> iter  20 value 361.798871
#> iter  30 value 360.507131
#> iter  40 value 360.388730
#> final  value 360.388034 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 369.839142
#> iter  20 value 357.303950
#> iter  30 value 356.102910
#> iter  40 value 356.075633
#> final  value 356.075168 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 371.837783
#> iter  20 value 360.206221
#> iter  30 value 359.213884
#> iter  40 value 359.197559
#> final  value 359.197181 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 369.841156
#> iter  20 value 357.306921
#> iter  30 value 356.106131
#> iter  40 value 356.078874
#> final  value 356.078407 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 372.348763
#> iter  20 value 359.726374
#> iter  30 value 358.652383
#> iter  40 value 358.605422
#> final  value 358.604910 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 374.363767
#> iter  20 value 362.559728
#> iter  30 value 361.700516
#> iter  40 value 361.617860
#> final  value 361.617614 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 372.350794
#> iter  20 value 359.729268
#> iter  30 value 358.655425
#> iter  40 value 358.608520
#> final  value 358.608009 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 371.820880
#> iter  20 value 356.681649
#> iter  30 value 355.393423
#> iter  40 value 355.269621
#> final  value 355.268692 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 373.796890
#> iter  20 value 359.616843
#> iter  30 value 358.560624
#> iter  40 value 358.499343
#> final  value 358.499018 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 371.822871
#> iter  20 value 356.684644
#> iter  30 value 355.396724
#> iter  40 value 355.272970
#> final  value 355.272038 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 366.113260
#> iter  20 value 354.801890
#> iter  30 value 353.764184
#> iter  40 value 353.708177
#> final  value 353.707915 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 368.185936
#> iter  20 value 357.746754
#> iter  30 value 356.845070
#> iter  40 value 356.791420
#> final  value 356.791344 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 366.115350
#> iter  20 value 354.804904
#> iter  30 value 353.767401
#> iter  40 value 353.711365
#> final  value 353.711103 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 371.295762
#> iter  20 value 357.558403
#> iter  30 value 355.786815
#> iter  40 value 355.650857
#> final  value 355.649580 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 373.365846
#> iter  20 value 360.761549
#> iter  30 value 359.398442
#> iter  40 value 359.329574
#> final  value 359.329346 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 371.297850
#> iter  20 value 357.561713
#> iter  30 value 355.790684
#> iter  40 value 355.654828
#> final  value 355.653553 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 367.448365
#> iter  20 value 355.694264
#> iter  30 value 354.549321
#> iter  40 value 354.513830
#> iter  40 value 354.513829
#> final  value 354.513775 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 369.523369
#> iter  20 value 358.493389
#> iter  30 value 357.609117
#> iter  40 value 357.576989
#> final  value 357.576895 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 367.450456
#> iter  20 value 355.697115
#> iter  30 value 354.552465
#> iter  40 value 354.516980
#> iter  40 value 354.516979
#> final  value 354.516924 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 372.980012
#> iter  20 value 359.385607
#> iter  30 value 358.277381
#> iter  40 value 358.192242
#> final  value 358.192054 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 374.964295
#> iter  20 value 362.339299
#> iter  30 value 361.380847
#> iter  40 value 361.344835
#> final  value 361.344655 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 372.982010
#> iter  20 value 359.388630
#> iter  30 value 358.280567
#> iter  40 value 358.195505
#> final  value 358.195316 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 373.148724
#> iter  20 value 358.683483
#> iter  30 value 357.506817
#> iter  40 value 357.407783
#> final  value 357.407614 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 375.161924
#> iter  20 value 361.698127
#> iter  30 value 360.695127
#> iter  40 value 360.632835
#> final  value 360.632526 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 373.150753
#> iter  20 value 358.686562
#> iter  30 value 357.510181
#> iter  40 value 357.411121
#> final  value 357.410952 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  48 (33 variable)
#> initial  value 833.162911 
#> iter  10 value 414.641193
#> iter  20 value 402.278808
#> iter  30 value 401.329882
#> iter  40 value 401.277606
#> final  value 401.276704 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex      married        age    havejob        educ
#> Formal     -2.571927  0.01603802 0.1004617753 -0.8335346  0.2070693  0.17887977
#> Informal   -2.637262 -0.08240285 0.0017005614 -0.5899114 -0.1280945 -0.01537467
#> L.Both     -3.346562 -0.08272813 0.0007447835 -1.0115396  0.2488088  0.31802075
#>               rural      region      income political.afl    networth
#> Formal    0.2873729 -0.28993596  0.72505922    0.21930816  0.23829653
#> Informal -0.3343724  0.14471252 -0.73309041   -0.22959659 -1.16969304
#> L.Both   -0.1768175 -0.02739693 -0.07475585   -0.06186197 -0.01215157
#> 
#> Residual Deviance: 802.5534 
#> AIC: 868.5534 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7285