Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 395.941779
#> iter  20 value 369.466050
#> iter  30 value 353.635072
#> iter  40 value 350.639784
#> final  value 350.607092 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 406.735175
#> iter  20 value 373.449454
#> iter  30 value 355.856610
#> iter  40 value 354.437964
#> final  value 354.436090 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 395.952834
#> iter  20 value 369.465128
#> iter  30 value 353.645430
#> iter  40 value 350.643062
#> final  value 350.611215 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 429.251487
#> iter  20 value 382.924452
#> iter  30 value 347.438967
#> iter  40 value 345.362568
#> final  value 345.360068 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 365.287593
#> iter  20 value 349.991249
#> iter  30 value 348.802446
#> iter  40 value 348.750141
#> final  value 348.749979 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 429.259264
#> iter  20 value 382.992617
#> iter  30 value 347.461694
#> iter  40 value 345.366128
#> final  value 345.363567 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 397.301220
#> iter  20 value 369.092699
#> iter  30 value 351.387873
#> iter  40 value 350.809383
#> final  value 350.809223 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 408.080389
#> iter  20 value 369.196560
#> iter  30 value 356.352097
#> iter  40 value 354.550823
#> final  value 354.519179 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 397.312268
#> iter  20 value 369.096258
#> iter  30 value 351.393118
#> iter  40 value 350.813386
#> final  value 350.813163 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 752.757838 
#> iter  10 value 407.983645
#> iter  20 value 369.646889
#> iter  30 value 351.466275
#> iter  40 value 349.875725
#> final  value 349.873769 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 752.757838 
#> iter  10 value 419.472916
#> iter  20 value 375.369073
#> iter  30 value 355.678713
#> iter  40 value 353.869988
#> final  value 353.861271 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 752.757838 
#> iter  10 value 407.995437
#> iter  20 value 369.650241
#> iter  30 value 351.463367
#> iter  40 value 349.880018
#> final  value 349.878087 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 394.557625
#> iter  20 value 370.878208
#> iter  30 value 355.935209
#> iter  40 value 352.446622
#> final  value 352.422299 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 405.532060
#> iter  20 value 372.220486
#> iter  30 value 357.517360
#> iter  40 value 356.369502
#> final  value 356.368336 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 394.568915
#> iter  20 value 370.881718
#> iter  30 value 355.946036
#> iter  40 value 352.450735
#> final  value 352.426527 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 404.824653
#> iter  20 value 371.843553
#> iter  30 value 351.154483
#> iter  40 value 348.549760
#> final  value 348.530588 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 416.679766
#> iter  20 value 380.068049
#> iter  30 value 355.226299
#> iter  40 value 352.644133
#> final  value 352.618734 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 404.836822
#> iter  20 value 371.850915
#> iter  30 value 351.153067
#> iter  40 value 348.555830
#> final  value 348.535038 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 403.507266
#> iter  20 value 365.510182
#> iter  30 value 352.421964
#> iter  40 value 351.313422
#> final  value 351.312369 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 413.740629
#> iter  20 value 369.573848
#> iter  30 value 356.006696
#> iter  40 value 354.872475
#> final  value 354.871373 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 403.517775
#> iter  20 value 365.511630
#> iter  30 value 352.425119
#> iter  40 value 351.317180
#> final  value 351.316129 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 397.358908
#> iter  20 value 368.053391
#> iter  30 value 346.272496
#> iter  40 value 343.611574
#> final  value 343.599796 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 408.597214
#> iter  20 value 367.974314
#> iter  30 value 348.356108
#> iter  40 value 347.639708
#> final  value 347.639081 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 397.370450
#> iter  20 value 368.056987
#> iter  30 value 346.273883
#> iter  40 value 343.616044
#> final  value 343.604172 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 394.283201
#> iter  20 value 362.732506
#> iter  30 value 349.345564
#> iter  40 value 348.035558
#> final  value 348.031163 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 404.880080
#> iter  20 value 366.660446
#> iter  30 value 353.121686
#> iter  40 value 352.391502
#> final  value 352.388307 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 394.294058
#> iter  20 value 362.738295
#> iter  30 value 349.348858
#> iter  40 value 348.040415
#> final  value 348.036033 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 405.662373
#> iter  20 value 370.082881
#> iter  30 value 354.753160
#> iter  40 value 353.404132
#> final  value 353.403998 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 416.087065
#> iter  20 value 375.896537
#> iter  30 value 357.853844
#> iter  40 value 357.128294
#> final  value 357.090028 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 405.673072
#> iter  20 value 370.089802
#> iter  30 value 354.755930
#> iter  40 value 353.408066
#> final  value 353.407931 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  48 (33 variable)
#> initial  value 833.162911 
#> iter  10 value 478.576245
#> iter  20 value 426.081248
#> iter  30 value 395.502639
#> iter  40 value 394.095036
#> final  value 394.093321 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex     married        age    havejob        educ
#> Formal     -2.564398 -0.08256392  0.03073382 -0.9830812  0.1325216 0.089789768
#> Informal   -2.830597 -0.03473112 -0.09093884 -0.6671989 -0.1243037 0.004238542
#> L.Both     -3.197716 -0.13896056 -0.26606661 -0.7819346  0.3563464 0.245169139
#>               rural     region    income political.afl   networth
#> Formal    0.4331606 -0.2865212  1.520020    0.24819458  0.3804303
#> Informal -0.3308053  0.2668837 -1.755510   -0.64856573 -0.9596667
#> L.Both   -0.3801607 -0.2365604  1.120561   -0.02077571  0.1531691
#> 
#> Residual Deviance: 788.1866 
#> AIC: 854.1866 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7227