Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
xvar <- c("sex", "married", "age", "havejob", "educ", "political.afl",
"rural", "region", "fin.intermdiaries", "fin.knowldge", "income")
BchMk.MLM <- MLM_Model(sample_data, c(xvar, "networth"), yvar )
#> + Fold01: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 390.189095
#> iter  20 value 366.164169
#> iter  30 value 362.638149
#> iter  40 value 353.143456
#> iter  50 value 351.871912
#> final  value 351.871868 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 404.690655
#> iter  20 value 379.458555
#> iter  30 value 359.553939
#> iter  40 value 356.500037
#> final  value 356.416071 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 390.204089
#> iter  20 value 366.178954
#> iter  30 value 362.647646
#> iter  40 value 353.168374
#> iter  50 value 351.877119
#> final  value 351.877034 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 419.881851
#> iter  20 value 389.894461
#> iter  30 value 363.165440
#> iter  40 value 360.187220
#> iter  50 value 360.116426
#> final  value 360.116420 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 432.264391
#> iter  20 value 391.558929
#> iter  30 value 367.008465
#> iter  40 value 363.814790
#> iter  50 value 363.790915
#> iter  50 value 363.790913
#> iter  50 value 363.790913
#> final  value 363.790913 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 419.894608
#> iter  20 value 389.902015
#> iter  30 value 363.162958
#> iter  40 value 360.228657
#> iter  50 value 360.120420
#> iter  50 value 360.120417
#> iter  50 value 360.120417
#> final  value 360.120417 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 409.973085
#> iter  20 value 382.059992
#> iter  30 value 355.717250
#> iter  40 value 352.548630
#> final  value 352.516572 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 423.549791
#> iter  20 value 386.523549
#> iter  30 value 359.338148
#> iter  40 value 356.988829
#> iter  50 value 356.299598
#> final  value 356.298575 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 409.987039
#> iter  20 value 382.065458
#> iter  30 value 355.707762
#> iter  40 value 352.552624
#> final  value 352.520669 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 418.227240
#> iter  20 value 376.373015
#> iter  30 value 357.468605
#> iter  40 value 355.255485
#> iter  50 value 355.215816
#> iter  50 value 355.215813
#> iter  50 value 355.215813
#> final  value 355.215813 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 430.992888
#> iter  20 value 384.158108
#> iter  30 value 364.990690
#> iter  40 value 359.479628
#> iter  50 value 358.961101
#> final  value 358.959507 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 418.240369
#> iter  20 value 376.383910
#> iter  30 value 357.488800
#> iter  40 value 355.275023
#> iter  50 value 355.235222
#> iter  60 value 355.223637
#> final  value 355.220426 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 493.217141
#> iter  20 value 386.128046
#> iter  30 value 359.191752
#> iter  40 value 354.113499
#> iter  50 value 354.077689
#> final  value 354.077675 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 500.590244
#> iter  20 value 388.846660
#> iter  30 value 361.946977
#> iter  40 value 357.686497
#> iter  50 value 357.597873
#> final  value 357.597862 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 748.598955 
#> iter  10 value 493.225491
#> iter  20 value 386.132872
#> iter  30 value 359.196275
#> iter  40 value 354.117718
#> iter  50 value 354.081463
#> final  value 354.081449 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 402.814905
#> iter  20 value 374.041126
#> iter  30 value 352.165910
#> iter  40 value 346.553149
#> iter  50 value 346.483916
#> iter  50 value 346.483915
#> iter  50 value 346.483915
#> final  value 346.483915 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 414.868285
#> iter  20 value 378.327828
#> iter  30 value 353.100507
#> iter  40 value 350.546340
#> iter  50 value 350.511182
#> iter  50 value 350.511182
#> iter  50 value 350.511182
#> final  value 350.511182 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 402.827235
#> iter  20 value 374.048168
#> iter  30 value 349.402414
#> iter  40 value 346.553442
#> iter  50 value 346.488239
#> iter  50 value 346.488237
#> iter  50 value 346.488237
#> final  value 346.488237 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 416.987548
#> iter  20 value 381.659211
#> iter  30 value 369.113737
#> iter  40 value 357.366430
#> iter  50 value 357.074680
#> iter  50 value 357.074678
#> iter  50 value 357.074678
#> final  value 357.074678 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 429.261716
#> iter  20 value 391.851807
#> iter  30 value 364.395028
#> iter  40 value 361.010257
#> iter  50 value 360.896872
#> iter  50 value 360.896870
#> iter  50 value 360.896870
#> final  value 360.896870 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 751.371544 
#> iter  10 value 417.000252
#> iter  20 value 381.660475
#> iter  30 value 369.161645
#> iter  40 value 357.369951
#> iter  50 value 357.078813
#> iter  50 value 357.078811
#> iter  50 value 357.078811
#> final  value 357.078811 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 492.434153
#> iter  20 value 383.744008
#> iter  30 value 360.665365
#> iter  40 value 356.050340
#> iter  50 value 355.926471
#> final  value 355.926437 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 500.628534
#> iter  20 value 389.608986
#> iter  30 value 365.120651
#> iter  40 value 359.750021
#> iter  50 value 359.685537
#> final  value 359.685527 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 492.443308
#> iter  20 value 383.749990
#> iter  30 value 360.670646
#> iter  40 value 356.054778
#> iter  50 value 355.930510
#> final  value 355.930469 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 374.801410
#> iter  20 value 359.934353
#> iter  30 value 358.557645
#> iter  40 value 358.413562
#> iter  50 value 358.410578
#> iter  50 value 358.410577
#> iter  50 value 358.410577
#> final  value 358.410577 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 376.829235
#> iter  20 value 362.585755
#> iter  30 value 361.741712
#> iter  40 value 361.618506
#> final  value 361.616820 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 374.803457
#> iter  20 value 359.937369
#> iter  30 value 358.560911
#> iter  40 value 358.416868
#> iter  50 value 358.413888
#> iter  50 value 358.413887
#> iter  50 value 358.413886
#> final  value 358.413886 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 493.880021
#> iter  20 value 380.761845
#> iter  30 value 359.092519
#> iter  40 value 353.833831
#> iter  50 value 353.725149
#> final  value 353.725099 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 502.173209
#> iter  20 value 388.370303
#> iter  30 value 365.086920
#> iter  40 value 357.966282
#> iter  50 value 357.881277
#> final  value 357.881271 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  56 (39 variable)
#> initial  value 749.985249 
#> iter  10 value 493.889168
#> iter  20 value 380.770070
#> iter  30 value 359.095571
#> iter  40 value 353.838586
#> iter  50 value 353.729758
#> final  value 353.729709 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  56 (39 variable)
#> initial  value 833.162911 
#> iter  10 value 416.113945
#> iter  20 value 401.363880
#> iter  30 value 400.332356
#> iter  40 value 400.105386
#> iter  50 value 399.949630
#> iter  50 value 399.949629
#> iter  50 value 399.949629
#> final  value 399.949629 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)          sex    married        age      havejob       educ
#> Formal     -2.659253 -0.021537862  0.1975061 -0.9094484  0.325197181 0.13310268
#> Informal   -2.619400  0.059575981 -0.1317913 -0.4214938 -0.001904514 0.08209539
#> L.Both     -3.435860  0.003430076  0.1096084 -0.9144233  0.465832201 0.34271215
#>          political.afl      rural      region fin.intermdiaries fin.knowldge
#> Formal       0.3531963  0.4231619 -0.21824917        0.02891838   0.10540702
#> Informal    -0.4044312 -0.3067345  0.15637728       -0.20220479   0.19984881
#> L.Both       0.1773511 -0.1550781 -0.03357011        0.15212475   0.02869293
#>              income   networth
#> Formal    0.7313618  0.4550824
#> Informal -2.1378193 -0.7227105
#> L.Both   -0.3554539 -0.1136251
#> 
#> Residual Deviance: 799.8993 
#> AIC: 877.8993 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7104