Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 368.044810
#> iter  20 value 352.645804
#> iter  30 value 351.726701
#> iter  40 value 351.622328
#> final  value 351.622296 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 370.058133
#> iter  20 value 355.661639
#> iter  30 value 354.891267
#> final  value 354.839920 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 368.046841
#> iter  20 value 352.648884
#> iter  30 value 351.730002
#> iter  40 value 351.625650
#> final  value 351.625617 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 368.688191
#> iter  20 value 356.582690
#> iter  30 value 355.095892
#> iter  40 value 355.012397
#> final  value 355.012033 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 370.781901
#> iter  20 value 359.825501
#> iter  30 value 358.303092
#> iter  40 value 358.221278
#> final  value 358.221048 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 368.690304
#> iter  20 value 356.585977
#> iter  30 value 355.099213
#> iter  40 value 355.015713
#> final  value 355.015347 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 442.967993
#> iter  20 value 391.032215
#> iter  30 value 367.014733
#> iter  40 value 363.498730
#> final  value 363.484767 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 453.198663
#> iter  20 value 393.054609
#> iter  30 value 369.144603
#> iter  40 value 366.526286
#> final  value 366.505790 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 442.978627
#> iter  20 value 391.035610
#> iter  30 value 367.029652
#> iter  40 value 363.503043
#> final  value 363.487882 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 368.749009
#> iter  20 value 355.495276
#> iter  30 value 354.467963
#> iter  40 value 354.422005
#> final  value 354.421492 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 370.835500
#> iter  20 value 358.462094
#> iter  30 value 357.678443
#> iter  40 value 357.613710
#> final  value 357.613500 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 368.751112
#> iter  20 value 355.498314
#> iter  30 value 354.471279
#> iter  40 value 354.425320
#> final  value 354.424805 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 369.196147
#> iter  20 value 357.704116
#> iter  30 value 356.932041
#> iter  40 value 356.926711
#> final  value 356.926697 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 371.331381
#> iter  20 value 360.628597
#> iter  30 value 359.993881
#> iter  40 value 359.988367
#> iter  40 value 359.988367
#> iter  40 value 359.988366
#> final  value 359.988366 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 369.198301
#> iter  20 value 357.707099
#> iter  30 value 356.935184
#> iter  40 value 356.929857
#> final  value 356.929843 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 369.419018
#> iter  20 value 356.799032
#> iter  30 value 355.846488
#> iter  40 value 355.764767
#> final  value 355.764536 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 371.467996
#> iter  20 value 359.556329
#> iter  30 value 358.790829
#> iter  40 value 358.731308
#> final  value 358.731141 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 369.421085
#> iter  20 value 356.801838
#> iter  30 value 355.849512
#> iter  40 value 355.767815
#> final  value 355.767584 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 376.162936
#> iter  20 value 362.055842
#> iter  30 value 361.266152
#> iter  40 value 361.122068
#> final  value 361.121581 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 378.180539
#> iter  20 value 365.039222
#> iter  30 value 364.437041
#> iter  40 value 364.327424
#> final  value 364.327346 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 376.164971
#> iter  20 value 362.058900
#> iter  30 value 361.269424
#> iter  40 value 361.125392
#> final  value 361.124906 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 506.509540
#> iter  20 value 395.345028
#> iter  30 value 363.204250
#> iter  40 value 358.883710
#> final  value 358.795652 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 373.002457
#> iter  20 value 363.061471
#> iter  30 value 361.877500
#> iter  40 value 361.811980
#> final  value 361.811937 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 506.519491
#> iter  20 value 395.351184
#> iter  30 value 363.218348
#> iter  40 value 358.886072
#> final  value 358.798762 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 369.308526
#> iter  20 value 357.413658
#> iter  30 value 356.375348
#> iter  40 value 356.358023
#> final  value 356.357897 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 371.455968
#> iter  20 value 360.537732
#> iter  30 value 359.632013
#> iter  40 value 359.611786
#> final  value 359.611771 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 369.310692
#> iter  20 value 357.416864
#> iter  30 value 356.378719
#> iter  40 value 356.361404
#> final  value 356.361278 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 361.657547
#> iter  20 value 347.756087
#> iter  30 value 346.095534
#> iter  40 value 346.034793
#> final  value 346.034657 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 363.791434
#> iter  20 value 351.278974
#> iter  30 value 349.508693
#> iter  40 value 349.428374
#> final  value 349.428329 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 361.659700
#> iter  20 value 347.759716
#> iter  30 value 346.099065
#> iter  40 value 346.038304
#> final  value 346.038168 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  48 (33 variable)
#> initial  value 833.162911 
#> iter  10 value 412.741955
#> iter  20 value 401.390499
#> iter  30 value 400.707489
#> iter  40 value 400.663683
#> final  value 400.663619 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex     married        age     havejob        educ
#> Formal     -2.656439 -0.07349265 -0.05855119 -0.8727169  0.30575150  0.05669595
#> Informal   -2.603015 -0.07688907 -0.12241407 -0.6130713 -0.07838638 -0.03961306
#> L.Both     -3.268219 -0.08148456 -0.24112902 -0.9540546  0.21407851  0.32094222
#>               rural       region     income political.afl   networth
#> Formal    0.5884186 -0.370318078  0.6594968   0.052691514  0.3446993
#> Informal -0.2628256  0.161975274 -0.7179809  -0.352911606 -1.0268452
#> L.Both   -0.3371748  0.009954277  0.4149467  -0.005085675  0.1320926
#> 
#> Residual Deviance: 801.3272 
#> AIC: 867.3272 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7651