Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 377.148544
#> iter  20 value 360.999041
#> iter  30 value 359.451535
#> iter  40 value 359.378475
#> final  value 359.378190 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 379.094273
#> iter  20 value 363.951477
#> iter  30 value 362.675791
#> iter  40 value 362.612479
#> final  value 362.612413 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 377.150505
#> iter  20 value 361.002059
#> iter  30 value 359.454847
#> iter  40 value 359.381811
#> final  value 359.381526 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 463.883647
#> iter  20 value 368.022515
#> iter  30 value 355.534328
#> iter  40 value 346.394906
#> final  value 346.388842 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 404.206333
#> iter  20 value 373.451013
#> iter  30 value 351.265774
#> iter  40 value 350.235816
#> final  value 350.234721 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 747.212661 
#> iter  10 value 463.893199
#> iter  20 value 368.034865
#> iter  30 value 355.524013
#> iter  40 value 346.398842
#> final  value 346.392939 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 408.305317
#> iter  20 value 383.063204
#> iter  30 value 359.973257
#> iter  40 value 357.668990
#> final  value 357.654304 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 419.676816
#> iter  20 value 385.241179
#> iter  30 value 362.736727
#> iter  40 value 361.261572
#> final  value 361.255073 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 408.316949
#> iter  20 value 383.057840
#> iter  30 value 359.995011
#> iter  40 value 357.660278
#> final  value 357.658160 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 401.225559
#> iter  20 value 375.608736
#> iter  30 value 358.342098
#> iter  40 value 354.234719
#> final  value 354.232989 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 413.272325
#> iter  20 value 380.724217
#> iter  30 value 359.088076
#> iter  40 value 358.031774
#> final  value 358.031760 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 401.237942
#> iter  20 value 375.618227
#> iter  30 value 358.310212
#> iter  40 value 354.238715
#> final  value 354.237073 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 407.729210
#> iter  20 value 376.053135
#> iter  30 value 357.526689
#> iter  40 value 356.368725
#> final  value 356.367546 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 418.563363
#> iter  20 value 378.643167
#> iter  30 value 360.418723
#> iter  40 value 359.971277
#> final  value 359.970661 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 407.740302
#> iter  20 value 376.062253
#> iter  30 value 357.532194
#> iter  40 value 356.372301
#> final  value 356.371357 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 752.757838 
#> iter  10 value 501.100463
#> iter  20 value 388.828551
#> iter  30 value 361.067915
#> iter  40 value 358.356478
#> final  value 358.354089 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 752.757838 
#> iter  10 value 510.172833
#> iter  20 value 392.408828
#> iter  30 value 365.961122
#> iter  40 value 361.745828
#> iter  50 value 361.737365
#> iter  50 value 361.737364
#> iter  50 value 361.737364
#> final  value 361.737364 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 752.757838 
#> iter  10 value 501.110546
#> iter  20 value 388.823066
#> iter  30 value 361.070718
#> iter  40 value 358.360015
#> final  value 358.357626 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 476.942655
#> iter  20 value 378.359919
#> iter  30 value 354.002049
#> iter  40 value 352.767584
#> final  value 352.764453 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 484.482103
#> iter  20 value 388.732840
#> iter  30 value 358.761957
#> iter  40 value 356.381266
#> final  value 356.375911 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 476.951093
#> iter  20 value 378.363513
#> iter  30 value 354.006357
#> iter  40 value 352.771432
#> final  value 352.768302 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 420.891436
#> iter  20 value 379.825955
#> iter  30 value 360.395604
#> iter  40 value 359.553675
#> final  value 359.553387 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 432.178954
#> iter  20 value 379.552043
#> iter  30 value 363.373658
#> iter  40 value 362.978742
#> final  value 362.977127 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 420.902996
#> iter  20 value 379.823576
#> iter  30 value 360.397445
#> iter  40 value 359.557313
#> final  value 359.557003 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 421.836265
#> iter  20 value 382.836451
#> iter  30 value 364.221211
#> iter  40 value 362.170646
#> final  value 362.167251 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 432.090981
#> iter  20 value 392.618038
#> iter  30 value 366.892162
#> iter  40 value 365.529887
#> final  value 365.507838 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 421.846763
#> iter  20 value 382.839370
#> iter  30 value 364.227355
#> iter  40 value 362.174252
#> final  value 362.170774 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 406.341333
#> iter  20 value 374.042677
#> iter  30 value 353.110327
#> iter  40 value 352.200305
#> final  value 352.174958 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 417.965453
#> iter  20 value 374.045406
#> iter  30 value 357.033766
#> iter  40 value 355.880576
#> final  value 355.864334 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 406.353235
#> iter  20 value 374.047023
#> iter  30 value 353.127023
#> iter  40 value 352.204241
#> final  value 352.178887 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  48 (33 variable)
#> initial  value 833.162911 
#> iter  10 value 492.303540
#> iter  20 value 440.917796
#> iter  30 value 406.496202
#> iter  40 value 400.793281
#> final  value 400.767421 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex    married        age     havejob        educ
#> Formal     -2.665319 -0.07006375  0.1019091 -1.0888487  0.07262497  0.11223643
#> Informal   -2.592090 -0.09035460 -0.1070482 -0.5646861 -0.01638885 -0.06150119
#> L.Both     -3.310459 -0.13066955 -0.2477353 -1.0630466  0.17197570  0.28657127
#>               rural      region     income political.afl   networth
#> Formal    0.4647567 -0.29778270  0.8405951     0.3284912  0.4848658
#> Informal -0.2231462  0.29053302 -1.6840056    -0.1961806 -0.9013424
#> L.Both   -0.4017999  0.06674012  0.5016706     0.2042903  0.1799053
#> 
#> Residual Deviance: 801.5348 
#> AIC: 867.5348 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7375