Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 455.926293
#> iter  20 value 361.315940
#> iter  30 value 343.296557
#> iter  40 value 342.665059
#> final  value 342.664745 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 464.517563
#> iter  20 value 366.939698
#> iter  30 value 347.195725
#> iter  40 value 346.629948
#> final  value 346.629726 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 455.935684
#> iter  20 value 361.322316
#> iter  30 value 343.304967
#> iter  40 value 342.669324
#> final  value 342.669007 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 428.164643
#> iter  20 value 364.446593
#> iter  30 value 342.362408
#> iter  40 value 341.795503
#> final  value 341.795282 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 437.053863
#> iter  20 value 366.906965
#> iter  30 value 346.036733
#> iter  40 value 345.715408
#> final  value 345.715291 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 428.174427
#> iter  20 value 364.458529
#> iter  30 value 342.365881
#> iter  40 value 341.799662
#> final  value 341.799452 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 472.596683
#> iter  20 value 378.137083
#> iter  30 value 349.128573
#> iter  40 value 347.690796
#> final  value 347.688927 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 482.939453
#> iter  20 value 381.410120
#> iter  30 value 352.185618
#> iter  40 value 351.196351
#> final  value 351.195000 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 472.607894
#> iter  20 value 378.145245
#> iter  30 value 349.132577
#> iter  40 value 347.694370
#> final  value 347.692589 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 489.780891
#> iter  20 value 379.812997
#> iter  30 value 350.646332
#> iter  40 value 349.617577
#> final  value 349.616288 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 499.269554
#> iter  20 value 382.199322
#> iter  30 value 354.257307
#> iter  40 value 353.083638
#> final  value 353.082162 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 489.791221
#> iter  20 value 379.807255
#> iter  30 value 350.652546
#> iter  40 value 349.621090
#> final  value 349.619908 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 752.757838 
#> iter  10 value 436.566157
#> iter  20 value 381.954482
#> iter  30 value 359.458029
#> iter  40 value 357.169934
#> final  value 357.164803 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 752.757838 
#> iter  10 value 440.748183
#> iter  20 value 380.672258
#> iter  30 value 361.924501
#> iter  40 value 360.671025
#> final  value 360.668136 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 752.757838 
#> iter  10 value 436.571370
#> iter  20 value 381.964012
#> iter  30 value 359.482174
#> iter  40 value 357.173669
#> final  value 357.168497 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 469.403256
#> iter  20 value 381.243572
#> iter  30 value 358.042460
#> iter  40 value 356.075700
#> final  value 356.074969 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 475.028676
#> iter  20 value 386.898360
#> iter  30 value 361.904182
#> iter  40 value 359.528407
#> final  value 359.498333 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 469.409947
#> iter  20 value 381.215344
#> iter  30 value 358.060569
#> iter  40 value 356.079210
#> final  value 356.078562 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 464.904202
#> iter  20 value 374.357241
#> iter  30 value 349.096651
#> iter  40 value 345.380248
#> final  value 345.379254 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 474.517654
#> iter  20 value 378.455048
#> iter  30 value 351.933793
#> iter  40 value 349.185913
#> final  value 349.183261 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 464.914569
#> iter  20 value 374.363972
#> iter  30 value 349.096008
#> iter  40 value 345.384231
#> final  value 345.383254 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 463.518843
#> iter  20 value 379.559188
#> iter  30 value 353.829098
#> iter  40 value 351.193722
#> final  value 351.166924 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 472.302216
#> iter  20 value 374.641644
#> iter  30 value 356.038295
#> iter  40 value 354.703160
#> final  value 354.700323 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 463.528458
#> iter  20 value 379.568851
#> iter  30 value 353.690891
#> iter  40 value 351.197629
#> final  value 351.170637 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 455.041030
#> iter  20 value 363.833605
#> iter  30 value 343.332974
#> iter  40 value 339.580382
#> final  value 339.575358 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 453.357844
#> iter  20 value 366.212169
#> iter  30 value 344.726434
#> iter  40 value 343.409726
#> final  value 343.407998 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 455.056420
#> iter  20 value 363.832122
#> iter  30 value 343.374954
#> iter  40 value 339.583996
#> final  value 339.579387 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 365.023406
#> iter  20 value 350.076974
#> iter  30 value 348.518042
#> iter  40 value 348.394502
#> final  value 348.394110 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 367.098266
#> iter  20 value 353.083322
#> iter  30 value 351.806826
#> iter  40 value 351.668762
#> final  value 351.668520 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 365.025499
#> iter  20 value 350.080048
#> iter  30 value 348.521416
#> iter  40 value 348.397876
#> final  value 348.397485 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  48 (33 variable)
#> initial  value 833.162911 
#> iter  10 value 541.354438
#> iter  20 value 430.788811
#> iter  30 value 394.118854
#> iter  40 value 391.886124
#> final  value 391.883817 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex     married        age     havejob        educ
#> Formal     -2.704254 -0.02773102  0.04746488 -0.8384138  0.39038598 0.196792717
#> Informal   -2.660242 -0.11303881 -0.17177793 -0.6597185 -0.11148283 0.006154965
#> L.Both     -3.225911  0.02270210 -0.26267417 -0.9632155  0.05034489 0.432333702
#>               rural     region     income political.afl   networth
#> Formal    0.4657667 -0.4959079  1.3780670     0.2506255  0.3305583
#> Informal -0.2967098  0.2914588 -1.4255243    -0.3274444 -1.0498772
#> L.Both   -0.3947566 -0.1042528  0.8770766     0.1218242  0.1629020
#> 
#> Residual Deviance: 783.7676 
#> AIC: 849.7676 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7087