Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 388.925509
#> iter  20 value 370.664811
#> iter  30 value 354.695388
#> iter  40 value 351.664904
#> final  value 351.659655 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 400.548416
#> iter  20 value 379.762588
#> iter  30 value 357.835158
#> iter  40 value 355.697311
#> final  value 355.690038 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 388.937413
#> iter  20 value 370.673812
#> iter  30 value 354.687108
#> iter  40 value 351.668382
#> final  value 351.664012 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 479.567381
#> iter  20 value 384.690629
#> iter  30 value 358.884110
#> iter  40 value 349.023167
#> final  value 349.007195 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 403.273789
#> iter  20 value 376.124905
#> iter  30 value 354.923065
#> iter  40 value 353.042520
#> final  value 353.028545 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 479.579888
#> iter  20 value 384.705024
#> iter  30 value 358.883660
#> iter  40 value 349.027235
#> final  value 349.011516 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 388.811927
#> iter  20 value 368.710411
#> iter  30 value 352.280238
#> iter  40 value 349.785321
#> final  value 349.780597 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 400.313907
#> iter  20 value 374.751349
#> iter  30 value 355.194477
#> iter  40 value 353.673863
#> final  value 353.656258 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 388.823717
#> iter  20 value 368.719072
#> iter  30 value 352.281201
#> iter  40 value 349.789395
#> final  value 349.784742 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 465.868063
#> iter  20 value 370.128297
#> iter  30 value 352.452656
#> iter  40 value 349.449330
#> final  value 349.447697 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 401.143687
#> iter  20 value 375.193333
#> iter  30 value 353.375087
#> iter  40 value 353.093513
#> final  value 353.093473 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 465.877428
#> iter  20 value 370.136305
#> iter  30 value 352.446619
#> iter  40 value 349.453224
#> final  value 349.451571 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 386.108244
#> iter  20 value 367.195660
#> iter  30 value 349.836867
#> iter  40 value 348.803859
#> final  value 348.802495 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 397.886435
#> iter  20 value 374.438783
#> iter  30 value 353.690130
#> iter  40 value 352.878900
#> final  value 352.878599 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 386.120288
#> iter  20 value 367.204800
#> iter  30 value 349.838950
#> iter  40 value 348.808179
#> final  value 348.806887 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 401.505597
#> iter  20 value 374.304980
#> iter  30 value 357.498513
#> iter  40 value 354.278539
#> final  value 354.256619 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 412.451827
#> iter  20 value 379.005415
#> iter  30 value 359.444339
#> iter  40 value 357.959310
#> final  value 357.956804 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 401.516768
#> iter  20 value 374.310690
#> iter  30 value 357.536367
#> iter  40 value 354.279315
#> final  value 354.260584 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 400.883413
#> iter  20 value 369.929926
#> iter  30 value 365.257957
#> iter  40 value 354.418398
#> final  value 354.243314 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 414.764784
#> iter  20 value 383.426397
#> iter  30 value 362.595954
#> iter  40 value 357.809776
#> final  value 357.770206 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 400.897439
#> iter  20 value 369.947003
#> iter  30 value 365.262890
#> iter  40 value 354.427055
#> final  value 354.246999 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 382.656017
#> iter  20 value 362.189356
#> iter  30 value 355.882953
#> iter  40 value 350.616213
#> final  value 350.603012 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 395.197639
#> iter  20 value 373.664731
#> iter  30 value 356.200290
#> iter  40 value 354.846189
#> final  value 354.844011 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 382.668880
#> iter  20 value 362.200723
#> iter  30 value 355.879449
#> iter  40 value 350.630294
#> final  value 350.607671 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 392.960255
#> iter  20 value 370.699086
#> iter  30 value 356.221725
#> iter  40 value 351.813503
#> final  value 351.797242 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 404.840466
#> iter  20 value 375.976104
#> iter  30 value 356.521113
#> iter  40 value 355.716891
#> final  value 355.712809 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 392.972427
#> iter  20 value 370.706949
#> iter  30 value 356.245370
#> iter  40 value 351.815538
#> final  value 351.801461 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 393.265863
#> iter  20 value 376.170773
#> iter  30 value 358.997645
#> iter  40 value 357.811029
#> final  value 357.810170 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 404.311531
#> iter  20 value 379.697272
#> iter  30 value 362.548332
#> iter  40 value 361.418967
#> final  value 361.415672 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 393.277172
#> iter  20 value 376.177617
#> iter  30 value 359.018293
#> iter  40 value 357.814824
#> final  value 357.813993 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 1e-04 on full training set
#> # weights:  48 (33 variable)
#> initial  value 833.162911 
#> iter  10 value 451.740825
#> iter  20 value 421.999076
#> iter  30 value 399.733788
#> iter  40 value 392.654698
#> final  value 392.627338 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex     married        age     havejob       educ
#> Formal     -2.832366 -0.02388029  0.18376409 -1.1178159  0.04496642  0.1919933
#> Informal   -2.880327 -0.12545184 -0.12071054 -0.6922472 -0.12187425 -0.0882684
#> L.Both     -3.456578 -0.05823638  0.01426225 -0.9141211  0.60639755  0.1999186
#>               rural      region     income political.afl    networth
#> Formal    0.7094825 -0.47109063  0.7814371    0.26595669  0.40066800
#> Informal -0.2570594  0.30587040 -2.3798384   -0.30535902 -1.10479302
#> L.Both   -0.2153652 -0.04578016  0.5259248    0.03242896  0.08041605
#> 
#> Residual Deviance: 785.2547 
#> AIC: 851.2547 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.6964