Multinominal Logistic Model

Multinominal Logistic Model

Usage

MLM_Model(Data, xvar, yvar)

Arguments

Data

The name of the Dataset.

xvar

X variables.

yvar

Y variable.

Value

The output from MLM_Model.

Details

Multi-nominal model is the generalized form of generalized logistic model and can be define as $$\pi_{i}^{h} = P(y_{i}^{h} = 1 | \bold{x}_{\bold{i}}^{h})$$ where \(h\) presents the class labels ("1-of-h") on the basis of an input vector \(x_j\), in our case \(x_j\) is loan types ("Formal Loan", "Informal Loan", "Both Loan", and "No Loan"). Furthermore,

\(y_{i}^h = 1\)if the weight w of \(x_j\) corresponds to belong a class and \(y_{i}^h=0\) otherwise. For \(i\) \(\in\) \(1,\ldots,h\) and the weight vectors w^i corresponds to class \(i\).

We set \({\bold{{w}}^{h}} = 0\) and the parameters to be learned are the weight vectors w^i for \(i\) \(\in\) \(1,\ldots,h-1\) . And the class probabilities must satisfy $$\sum_{i=1}^{h} P(y_{i}^{h} = 1|\bold{x}_{\bold{i}}^{h}, \bold{w}) = 1.$$

Examples

yvar <- c("Loan.Type")
sample_data <- sample_data[c(1:750),]
m2.xvar0 <- c("sex", "married", "age", "havejob", "educ", "rural", "region","income")
BchMk.MLM <- MLM_Model(sample_data, c(m2.xvar0, "political.afl", "networth"), yvar)
#> + Fold01: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 388.926836
#> iter  20 value 367.642931
#> iter  30 value 347.546590
#> iter  40 value 344.906420
#> final  value 344.890839 
#> converged
#> - Fold01: decay=0e+00 
#> + Fold01: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 402.239469
#> iter  20 value 372.625745
#> iter  30 value 349.472900
#> iter  40 value 348.800422
#> final  value 348.798463 
#> converged
#> - Fold01: decay=1e-01 
#> + Fold01: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 388.940461
#> iter  20 value 367.652933
#> iter  30 value 347.545625
#> iter  40 value 344.910269
#> final  value 344.895043 
#> converged
#> - Fold01: decay=1e-04 
#> + Fold02: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 402.807433
#> iter  20 value 377.130381
#> iter  30 value 356.526761
#> iter  40 value 354.877016
#> final  value 354.875708 
#> converged
#> - Fold02: decay=0e+00 
#> + Fold02: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 414.636008
#> iter  20 value 375.400015
#> iter  30 value 359.090641
#> iter  40 value 358.478622
#> final  value 358.476296 
#> converged
#> - Fold02: decay=1e-01 
#> + Fold02: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 402.819527
#> iter  20 value 377.132927
#> iter  30 value 356.528663
#> iter  40 value 354.880844
#> final  value 354.879521 
#> converged
#> - Fold02: decay=1e-04 
#> + Fold03: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 390.069311
#> iter  20 value 364.862514
#> iter  30 value 346.235865
#> iter  40 value 344.136663
#> final  value 344.135452 
#> converged
#> - Fold03: decay=0e+00 
#> + Fold03: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 403.433967
#> iter  20 value 371.117837
#> iter  30 value 349.150411
#> iter  40 value 348.369327
#> final  value 348.366460 
#> converged
#> - Fold03: decay=1e-01 
#> + Fold03: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 390.083015
#> iter  20 value 364.870623
#> iter  30 value 346.232404
#> iter  40 value 344.141204
#> final  value 344.140035 
#> converged
#> - Fold03: decay=1e-04 
#> + Fold04: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 406.102830
#> iter  20 value 377.513732
#> iter  30 value 355.497685
#> iter  40 value 354.720679
#> final  value 354.720374 
#> converged
#> - Fold04: decay=0e+00 
#> + Fold04: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 417.840971
#> iter  20 value 381.484193
#> iter  30 value 359.439795
#> iter  40 value 358.273727
#> final  value 358.232905 
#> converged
#> - Fold04: decay=1e-01 
#> + Fold04: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 406.114860
#> iter  20 value 377.513981
#> iter  30 value 355.505066
#> iter  40 value 354.724918
#> final  value 354.724089 
#> converged
#> - Fold04: decay=1e-04 
#> + Fold05: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 405.789536
#> iter  20 value 382.333017
#> iter  30 value 356.163695
#> iter  40 value 353.832500
#> final  value 353.807101 
#> converged
#> - Fold05: decay=0e+00 
#> + Fold05: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 418.317463
#> iter  20 value 383.041169
#> iter  30 value 359.251646
#> iter  40 value 357.323633
#> final  value 357.321020 
#> converged
#> - Fold05: decay=1e-01 
#> + Fold05: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 405.802409
#> iter  20 value 382.343216
#> iter  30 value 356.156998
#> iter  40 value 353.835983
#> final  value 353.810811 
#> converged
#> - Fold05: decay=1e-04 
#> + Fold06: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 366.675407
#> iter  20 value 353.149729
#> iter  30 value 351.709268
#> iter  40 value 351.637180
#> final  value 351.636608 
#> converged
#> - Fold06: decay=0e+00 
#> + Fold06: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 368.736757
#> iter  20 value 356.287171
#> iter  30 value 355.072874
#> iter  40 value 355.020725
#> final  value 355.020470 
#> converged
#> - Fold06: decay=1e-01 
#> + Fold06: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 366.677484
#> iter  20 value 353.152945
#> iter  30 value 351.712804
#> iter  40 value 351.640729
#> final  value 351.640156 
#> converged
#> - Fold06: decay=1e-04 
#> + Fold07: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 403.075130
#> iter  20 value 375.755175
#> iter  30 value 353.797224
#> iter  40 value 352.299879
#> final  value 352.228595 
#> converged
#> - Fold07: decay=0e+00 
#> + Fold07: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 415.826272
#> iter  20 value 375.180801
#> iter  30 value 356.287268
#> iter  40 value 355.803375
#> final  value 355.802945 
#> converged
#> - Fold07: decay=1e-01 
#> + Fold07: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 403.088222
#> iter  20 value 375.757577
#> iter  30 value 353.750510
#> iter  40 value 352.299841
#> final  value 352.232398 
#> converged
#> - Fold07: decay=1e-04 
#> + Fold08: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 485.326178
#> iter  20 value 355.101900
#> iter  30 value 334.489748
#> iter  40 value 333.703722
#> final  value 333.703457 
#> converged
#> - Fold08: decay=0e+00 
#> + Fold08: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 499.707524
#> iter  20 value 364.441475
#> iter  30 value 342.910096
#> iter  40 value 338.318152
#> final  value 338.309062 
#> converged
#> - Fold08: decay=1e-01 
#> + Fold08: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 751.371544 
#> iter  10 value 485.341797
#> iter  20 value 355.118220
#> iter  30 value 334.512901
#> iter  40 value 333.727671
#> final  value 333.727494 
#> converged
#> - Fold08: decay=1e-04 
#> + Fold09: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 490.525834
#> iter  20 value 373.430009
#> iter  30 value 355.606242
#> iter  40 value 349.291607
#> final  value 349.286979 
#> converged
#> - Fold09: decay=0e+00 
#> + Fold09: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 500.434566
#> iter  20 value 382.678694
#> iter  30 value 356.528062
#> iter  40 value 353.068870
#> final  value 353.063769 
#> converged
#> - Fold09: decay=1e-01 
#> + Fold09: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 749.985249 
#> iter  10 value 490.536882
#> iter  20 value 373.440407
#> iter  30 value 355.604883
#> iter  40 value 349.295671
#> final  value 349.291020 
#> converged
#> - Fold09: decay=1e-04 
#> + Fold10: decay=0e+00 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 480.133064
#> iter  20 value 370.624176
#> iter  30 value 346.570262
#> iter  40 value 345.553425
#> final  value 345.553129 
#> converged
#> - Fold10: decay=0e+00 
#> + Fold10: decay=1e-01 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 489.220123
#> iter  20 value 377.370958
#> iter  30 value 350.029482
#> iter  40 value 349.325704
#> final  value 349.324067 
#> converged
#> - Fold10: decay=1e-01 
#> + Fold10: decay=1e-04 
#> # weights:  48 (33 variable)
#> initial  value 748.598955 
#> iter  10 value 480.143145
#> iter  20 value 370.623662
#> iter  30 value 346.572125
#> iter  40 value 345.557411
#> final  value 345.557120 
#> converged
#> - Fold10: decay=1e-04 
#> Aggregating results
#> Selecting tuning parameters
#> Fitting decay = 0.1 on full training set
#> # weights:  48 (33 variable)
#> initial  value 833.162911 
#> iter  10 value 411.555025
#> iter  20 value 394.373608
#> iter  30 value 393.227925
#> iter  40 value 393.133919
#> final  value 393.119353 
#> converged
BchMk.MLM$finalModel
#> Call:
#> nnet::multinom(formula = .outcome ~ ., data = dat, decay = param$decay)
#> 
#> Coefficients:
#>          (Intercept)         sex      married        age    havejob       educ
#> Formal     -2.716729 -0.16440081  0.191640083 -0.9762633 0.23471937  0.3616275
#> Informal   -2.806666 -0.11167512 -0.145228929 -0.5858354 0.07399396 -0.1193460
#> L.Both     -3.374527 -0.04331957 -0.000640068 -1.0415592 0.39821284  0.3058681
#>               rural      region     income political.afl   networth
#> Formal    0.3740464 -0.27186163  0.8067193     0.2651930  0.4160833
#> Informal -0.2237758  0.25646810 -1.1302050    -0.2362583 -1.6190411
#> L.Both   -0.1427601  0.09421892  0.3491207     0.1036139  0.1473135
#> 
#> Residual Deviance: 786.2387 
#> AIC: 852.2387 
BchMk.MLM$Roc$auc
#> Multi-class area under the curve: 0.7075